refactor out duplicated search
